feat: update search popup
